Just a commit to get things committed really. Tried to implement the old classes: LuceneIndexService, LuceneFulltextIndexService and so on with the new index framework underneath.
git-svn-id: https://svn.neo4j.org/components@4491 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ index/branches/new-index-api/src/test/java/org/neo4j/index/lucene/TestLuceneIndexingService.java
@@ -106,7 +106,7 @@ public class TestLuceneIndexingService extends Neo4jWithIndexTestCase
         assertNull( ( (LuceneIndexService) index() ).getEnabledCacheSize( key ) );
         ( ( LuceneIndexService ) index() ).enableCache( key, 1000 );
         assertEquals( ( Integer ) 1000,
-                ( (LuceneIndexService) index() ).getEnabledCacheSize( key ) );
+                ( Integer ) ( (LuceneIndexService) index() ).getEnabledCacheSize( key ) );
         Node node1 = graphDb().createNode();
         index().index( node1, key, value );
         index().getNodes( key, value );
